requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
datasets==2.7.1
dill==0.3.6
docker-pycreds==0.4.0
filelock==3.8.2
fire==0.4.0
frozenlist==1.3.3
fsspec==2022.11.0
gensim==4.2.0
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
keybert==0.7.0
keytotext==2.3.2
lightning-utilities==0.3.0
multidict==6.0.3
multiprocess==0.70.14
nltk==3.7
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.2
pathtools==0.1.2
Pillow==9.3.0
progressbar==2.5
promise==2.3
protobuf==3.20.1
psutil==5.9.4
pyarrow==10.0.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytorch-lightning==1.8.3.post1
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
simple-image-download==0.4
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
soupsieve==2.3.2.post1
tensorboardX==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.6
wikipedia==1.4.0
xxhash==3.1.0
yarl==1.8.2
